Refactor: Centralize Stripe customer logic
This commit centralizes the Stripe customer creation and management logic into a new shared service, `StripeCustomerService`.

Previously, both the `payment` and `subscription` modules contained their own logic for interacting with Stripe's customer APIs. This resulted in duplicated code and a potential for inconsistencies.

This refactoring moves the `StripeCustomerService` from the `payment` module to the `shared/stripe` module, making it accessible to all parts of the application. The `PaymentService` and `StripeClientImpl` (used by `SubscriptionService`) have been updated to use this new shared service. This ensures that all Stripe customer operations are handled by a single, consistent implementation, reducing code duplication and improving maintainability.
