+++ backend/src/main/java/com/platform/shared/stripe/StripeClient.java
@@ -1,7 +1,6 @@
 package com.platform.shared.stripe;
 
-import java.util.UUID;
-
+import com.platform.user.internal.Organization;
 import com.stripe.exception.StripeException;
 import com.stripe.model.PaymentMethod;
 import com.stripe.model.Subscription;
@@ -9,7 +8,7 @@ import com.stripe.param.SubscriptionCreateParams;
 import com.stripe.param.SubscriptionUpdateParams;
 
 public interface StripeClient {
-  String getOrCreateCustomer(UUID organizationId, String organizationName) throws StripeException;
+  String getOrCreateCustomer(Organization organization) throws StripeException;
 
   PaymentMethod retrievePaymentMethod(String paymentMethodId) throws StripeException;
