+++ backend/src/main/java/com/platform/shared/stripe/StripeClientImpl.java
@@ -1,18 +1,13 @@
 package com.platform.shared.stripe;
 
-import java.util.UUID;
-
+import com.platform.user.internal.Organization;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import com.stripe.exception.StripeException;
-import com.stripe.model.Customer;
-import com.stripe.model.CustomerSearchResult;
 import com.stripe.model.PaymentMethod;
 import com.stripe.model.Subscription;
 import com.stripe.net.RequestOptions;
-import com.stripe.param.CustomerCreateParams;
-import com.stripe.param.CustomerSearchParams;
 import com.stripe.param.PaymentMethodAttachParams;
 import com.stripe.param.SubscriptionCreateParams;
 import com.stripe.param.SubscriptionUpdateParams;
@@ -21,33 +16,18 @@ import com.stripe.param.SubscriptionUpdateParams;
 public class StripeClientImpl implements StripeClient {
 
   private final RequestOptions requestOptions;
+  private final StripeCustomerService stripeCustomerService;
 
-  public StripeClientImpl(@Value("${stripe.secret-key}") String apiKey) {
+  public StripeClientImpl(
+      @Value("${stripe.secret-key}") String apiKey,
+      StripeCustomerService stripeCustomerService) {
     this.requestOptions = RequestOptions.builder().setApiKey(apiKey).build();
+    this.stripeCustomerService = stripeCustomerService;
   }
 
   @Override
-  public String getOrCreateCustomer(UUID organizationId, String organizationName)
-      throws StripeException {
-    CustomerSearchParams searchParams =
-        CustomerSearchParams.builder()
-            .setQuery("metadata['organization_id']:'" + organizationId + "'")
-            .build();
-
-    CustomerSearchResult searchResult = Customer.search(searchParams, requestOptions);
-    if (!searchResult.getData().isEmpty()) {
-      return searchResult.getData().get(0).getId();
-    }
-
-    CustomerCreateParams params =
-        CustomerCreateParams.builder()
-            .setName(organizationName)
-            .setDescription("Customer for organization: " + organizationName)
-            .putMetadata("organization_id", organizationId.toString())
-            .build();
-
-    Customer customer = Customer.create(params, requestOptions);
-    return customer.getId();
+  public String getOrCreateCustomer(Organization organization) throws StripeException {
+    return stripeCustomerService.getOrCreateCustomer(organization);
   }
 
   @Override
